requirements.txt
tensorflow>=2.16.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
flask>=2.0.0
folium>=0.14.0
requests>=2.25.0
tqdm>=4.60.0
geopy>=2.2.0